Workaround TF restriction on creating many S3 objs
In short, `count` cannot interpolate a resource attribute.

This means that there is no way to have Terraform discover the number of
files to store on S3, and then store them all - we need to write out a
`resource` for each of them.

I still don't want to do that manually, though, so with this commit the
static post generator will also create the TF configuration for each
post.

Unfortunately, this requires two `terraform plan`s to take effect.
(Presumably Terraform loads all the configuration into memory before
executing any of it - and indeed why would it monitor for additional
files.)